reorganize TransactionInputSerializer tests into new file
