+++ tests/Transaction/TransactionInputTest.php
@@ -4,24 +4,12 @@ namespace BitWasp\Bitcoin\Tests\Transaction;
 
 use BitWasp\Bitcoin\Script\Script;
 use BitWasp\Bitcoin\Script\ScriptFactory;
-use BitWasp\Bitcoin\Serializer\Transaction\OutPointSerializer;
-use BitWasp\Bitcoin\Serializer\Transaction\TransactionInputSerializer;
 use BitWasp\Bitcoin\Transaction\OutPoint;
 use BitWasp\Bitcoin\Transaction\TransactionInput;
 use BitWasp\Buffertools\Buffer;
 
 class TransactionInputTest extends \PHPUnit_Framework_TestCase
 {
-    /**
-     * @var string
-     */
-    private $baseType = 'BitWasp\Bitcoin\Transaction\TransactionInput';
-
-    /**
-     * @var string
-     */
-    private $scriptType = 'BitWasp\Bitcoin\Script\Script';
-
     public function testGetSequence()
     {
         // test default
@@ -56,7 +44,7 @@ class TransactionInputTest extends \PHPUnit_Framework_TestCase
         $script = new Script(Buffer::hex('41'));
         $in = new TransactionInput(new OutPoint(Buffer::hex('7f8e94bdf85de933d5417145e4b76926777fa2a2d8fe15b684cfd835f43b8b33', 32), '0'), $script);
 
-        $this->assertInstanceOf($this->scriptType, $in->getScript());
+        $this->assertInstanceOf(Script::class, $in->getScript());
         $this->assertEquals($script, $in->getScript());
     }
 
@@ -88,21 +76,5 @@ class TransactionInputTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($in1->equals($inBadOut));
         $this->assertFalse($in1->equals($inBadScript));
         $this->assertFalse($in1->equals($inBadSeq));
-    }
-
-    public function testFromParser()
-    {
-        $buffer = '62442ea8de9ee6cc2dd7d76dfc4523910eb2e3bd4b202d376910de700f63bf4b000000008b48304502207db5ea602fe2e9f8e70bfc68b7f468d68910d2ff4ac50294fc80109e254f317f022100a68a66f23406fdfd93025c28ffef4e79260283335ce39a4e8d0b52c5ee41913b014104f8de51f3b278225c0fe74a856ea2481e9ad4c9385fc10cefadaa4357ecd2c4d29904902d10e376546500c127f65d0de35b6215d49dd1ef6c67e6cdd5e781ef22ffffffff';
-        $s = new TransactionInputSerializer(new OutPointSerializer());
-        $in = $s->parse($buffer);
-        $this->assertInstanceOf($this->baseType, $in);
-    }
-
-    public function testSerialize()
-    {
-        $hex = '62442ea8de9ee6cc2dd7d76dfc4523910eb2e3bd4b202d376910de700f63bf4b000000008b48304502207db5ea602fe2e9f8e70bfc68b7f468d68910d2ff4ac50294fc80109e254f317f022100a68a66f23406fdfd93025c28ffef4e79260283335ce39a4e8d0b52c5ee41913b014104f8de51f3b278225c0fe74a856ea2481e9ad4c9385fc10cefadaa4357ecd2c4d29904902d10e376546500c127f65d0de35b6215d49dd1ef6c67e6cdd5e781ef22ffffffff';
-        $s = new TransactionInputSerializer(new OutPointSerializer());
-        $in = $s->parse($hex);
-        $this->assertEquals($hex, $in->getBuffer()->getHex());
     }
 }
